requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-chroma
chromadb
sentence-transformers
rank-bm25
pypdf
python-dotenv
pydantic
streamlit
requests
google-cloud-storage
google-cloud-bigquery==3.41.0
